+++ backend/HR System/EmployeeService.Infrastructure/Persistence/HolidayRequest/HolidayRequestRepository.cs
@@ -24,22 +24,32 @@ namespace EmployeeService.Infrastructure.Persistence.HolidayRequest
             return savedEntity.Entity;
         }
 
-        public async Task<Core.Entities.HolidayRequest?> DeleteHolidayRequestAsync(Guid id, CancellationToken cancellationToken = default)
+        public async Task<bool> DeleteHolidayRequestAsync(Guid id, CancellationToken cancellationToken = default)
         {
             Core.Entities.HolidayRequest deletedEntity = await _context.HolidayRequests.FindAsync(id);
+            if (deletedEntity is null)
+            {
+                return false;
+            }
             _context.HolidayRequests.Remove(deletedEntity);
             await _context.SaveChangesAsync(cancellationToken);
-            return deletedEntity;
+            return true;
         }
 
-        public async Task<IEnumerable<Core.Entities.HolidayRequest>> GetAllHolidayRequestsAsync(int page, CancellationToken cancellationToken = default)
+        public async Task<IEnumerable<Core.Entities.HolidayRequest>> GetAllHolidayRequestsAsync(int page, int items, CancellationToken cancellationToken = default)
         {
             var holidayRequests = await _context.HolidayRequests.OrderBy(x => x.Id)
             .Include(x => x.Sender)
-            .Skip((page - 1) * 10)
-            .Take(10)
+            .Skip((page - 1) * items)
+            .Take(items)
             .ToListAsync(cancellationToken);
             return holidayRequests;
+        }
+
+        public async Task<int> GetTotalPagesAsync(int page, int items, CancellationToken cancellationToken = default)
+        {
+            var count = await _context.HolidayRequests.CountAsync(cancellationToken);
+            return (int)Math.Ceiling((double)(count / items));
         }
 
         public async Task<Core.Entities.HolidayRequest?> GetHolidayRequestByIdAsync(Guid id, CancellationToken cancellationToken = default)
@@ -52,6 +62,11 @@ namespace EmployeeService.Infrastructure.Persistence.HolidayRequest
             _context.HolidayRequests.Update(holidayRequest);
             await _context.SaveChangesAsync(cancellationToken);
             return holidayRequest;
+        }
+
+        public async Task<bool> CheckHolidayRequestExistenceAsync(Guid senderId, DateTime start, DateTime end, CancellationToken cancellationToken = default)
+        {
+            return await _context.HolidayRequests.AsNoTracking().FirstOrDefaultAsync(x => x.SenderId == senderId && x.Start.Date == start.Date && x.End.Date == end.Date) is not null;
         }
     }
 }
